requirements.txt

streamlit
google-generativeai
moviepy
pydub
assemblyai
